requirements.txt
Django==5.1.4
Pillow==11.0.0
torch==2.6.0
diffusers==0.31.0
transformers==4.47.1
accelerate==1.2.1
safetensors==0.4.5
